Consolidate input components
