+++ client/src/Forms/SeriesForm.tsx
@@ -27,8 +27,8 @@ import { useSnackbar } from 'notistack';
 import React, { ReactElement, useEffect, useState } from 'react';
 import * as Yup from 'yup';
 import { GenericError, NetworkError } from '../components/ErrorSnackbars';
+import { SeriesAutocomplete } from '../components/Inputs/SeriesAutocomplete';
 import { SourceAutoComplete } from '../components/Inputs/SourceAutoComplete';
-import { SeriesAutocomplete } from '../components/SeriesAutocomplete';
 import { FormLoading } from '../components/Skeletons/FormLoading';
 import { Series, Type } from '../gql/documents';
 import {
