+++ client/src/components/Dashboard/AdminCard.tsx
@@ -13,8 +13,8 @@ import React, { useState } from 'react';
 import { File } from '../../gql/documents';
 import { useCatalogStatisticsQuery } from '../../gql/queries';
 import { renderDuration } from '../../utils/enumRender';
-import { CatalogDataExportButton } from '../CatalogDataExportButton';
-import { FileRecordsExportButton } from '../FileRecordsExportButton';
+import { CatalogDataExportButton } from '../Inputs/CatalogDataExportButton';
+import { FileRecordsExportButton } from '../Inputs/FileRecordsExportButton';
 import { CodecChart } from './CodecChart';
 import { ResolutionChart } from './ResolutionChart';
 import { SourceChart } from './SourceChart';
