+++ client/src/components/Inputs/CatalogDataExportButton.tsx
@@ -9,7 +9,9 @@ import { teal } from '@material-ui/core/colors';
 import { saveAs } from 'file-saver';
 import { useSnackbar } from 'notistack';
 import React from 'react';
-import { useFileExportLazyQuery } from '../gql/queries';
+import { Series } from '../../gql/documents';
+import { useExportDataLazyQuery } from '../../gql/queries';
+import { generateExcel } from '../../utils/excel';
 
 type Props = {
   setLoading: React.Dispatch<React.SetStateAction<boolean>>;
@@ -32,47 +34,34 @@ const useStyles = makeStyles((theme: Theme) =>
   })
 );
 
-export const FileRecordsExportButton = (props: Props) => {
+export const CatalogDataExportButton = (props: Props) => {
   const classes = useStyles();
   const { setLoading, loading: anyLoading } = props;
   const { enqueueSnackbar, closeSnackbar } = useSnackbar();
 
-  const [fileExportQuery] = useFileExportLazyQuery({
+  const [exportDataQuery] = useExportDataLazyQuery({
     fetchPolicy: 'no-cache',
     onCompleted: (data) => {
-      const content = data.files
-        ?.map((file) => file?.path)
-        .filter(Boolean)
-        .reduce((previous, current) => {
-          return previous + `"${current}",\r\n`;
-        }, '');
-      closeSnackbar('file-generation');
-      if (content) {
-        const file = new File([content], 'fileList.csv', {
-          type: 'text/csv,charset=utf-8',
+      const file = generateExcel(data.allSeries as Series[])
+        .then((file) => saveAs(file, 'Exported Data.xlsx'))
+        .then(() => {
+          closeSnackbar('file-generation');
+          enqueueSnackbar(`Generated file.`, {
+            key: 'file-generation-success',
+            variant: 'success',
+          });
+          setLoading(false);
         });
-        enqueueSnackbar(`Generated file.`, {
-          key: 'file-generation-success',
-          variant: 'success',
-        });
-        saveAs(file);
-      } else {
-        enqueueSnackbar(`Something went wrong. Please try again.`, {
-          key: 'file-generation',
-          variant: 'warning',
-        });
-      }
-      setLoading(false);
     },
   });
 
-  const handleClick = () => {
+  const handleClick = async () => {
     setLoading(true);
     enqueueSnackbar(`Generating file... Please wait...`, {
       key: 'file-generation',
       autoHideDuration: null,
     });
-    fileExportQuery();
+    exportDataQuery();
   };
 
   return (
@@ -82,7 +71,7 @@ export const FileRecordsExportButton = (props: Props) => {
         disabled={anyLoading}
         variant="contained"
       >
-        Export File List CSV
+        Export Data
       </Button>
       {anyLoading && (
         <CircularProgress size={24} className={classes.buttonProgress} />
